+++ CHANGELOG.md
@@ -9,7 +9,11 @@ Welcome to the Github! If you are looking for the most recent stable version, th
 # | Repurposed Structures changelog |
 
 
-## (V.1.8.8 Changes) (1.16.5 Minecraft)
+## (V.1.9.0 Changes) (1.16.5 Minecraft)
+
+##### Loot Tables:
+- Added code that now will import modded items from vanilla structure's loot tables into Repurposed Structures's own loot tables. 
+(Snowy Pyramid gets all modded items that vanilla Desert Temple can have) 
 
 ##### Language:
 - Fixed some Spanish translations being cut-off.